fix(tests): use C_LOC(array(1)) consistently and deduplicate Windows Unicode helpers
Replace bare C_LOC(array) with C_LOC(array(1)) across parallel Fortran
tests for portability across Fortran 2003 compilers that require an array
element (not a whole-array designator) as the C_LOC argument.

Also extract the duplicate Windows UTF-8 filesystem helpers (utf8_to_wide,
utf8_mkdir, utf8_unlink, utf8_rmdir) from test_unicode.c and
pcgns_unicode_test.c into a single shared header utils_unicode_win.h,
adding an integer-overflow guard on the allocation size while there.
